DB 포스트 테스트 & 광고 추가

+++ src/components/Tag/Tag.jsx
@@ -3,7 +3,7 @@ import "./Tag.scss";
 
 const Tag = ({ name, onRemove }) => {
   const removeTagOnClick = () => {
-    onRemove(name);
+    if (onRemove) onRemove(name);
   };
 
   return (
